[Fix] #448 - LeftAlignedCollectionViewFlowLayout 파일 분리
